# remote-externalities: 'instant' snapshots, threading refactor, better progress logs