--- a/primitives/state-machine/src/testing.rs
+++ b/primitives/state-machine/src/testing.rs
@@ -117,12 +117,7 @@ where
 		}
 	}
 
-	/// Returns a mutable reference to the overlayed changes.
-	pub fn overlayed_changes_mut(&mut self) -> &mut OverlayedChanges {
-		&mut self.overlay
-	}
-
-	/// Returns an immutable reference to the overlayed changes.
+	/// Returns the overlayed changes.
 	pub fn overlayed_changes(&self) -> &OverlayedChanges {
 		&self.overlay
 	}