## Instant(ish) snapshot loading
  - The existing snapshots contain key / values without the trie. Loading the key/values can take a while even when batch inserting because the trie needs to be re-computed.
  - In this PR, the snapshot was refactored to contain the raw trie itself, and be loaded directly into the underlying TrieBackend storage that backs TestExternalties

### Snapshot load performance (Macbook Pro M2 Max)

| Chain                       | Build   |  Before  | After  | Improvement |
|----------------------------|---------|----------|---------|-------------|
| Kusama | debug   |  12.13s  |  0.88s  |    92.75%   |
| Kusama | release |   1.93s  |  0.18s  |    90.67%   |
| Polkadot | debug   |  59.08s  |  2.66s  |    95.50%   |
| Polkadot | release |   8.25s  |  0.44s  |    94.67%   |

## Removes threading
  - Uses async/await futures for parallel requests instead of threads
  - Closes https://github.com/paritytech/substrate/issues/14006
  - It makes loading state from a live node a *little bit* slower because it waits until all data has been downloaded before loading it into the trie. However, the degradation is small, and it should be encouraged to use snapshots instead when performance is required especially now that we have much faster snapshot loading.
  - Bonus: simplifies a lot of complex code 

## Better progress logs
  - Spinner for operations where the eta cannot be estimated
  - Progress bar for operations where the eta can be estimated
  - Partial https://github.com/paritytech/substrate/issues/13975

### Kusama `on_runtime_upgrade` `live` logs

https://user-images.githubusercontent.com/16665596/235869599-6d50992a-15c7-4f99-b27b-a95d4319b87f.mp4

## TODO Check pulling data from these remotes works: 
- [x] `wss://polkadot-try-runtime-node.parity-chains.parity.io:443`
- [x] `wss://kusama-try-runtime-node.parity-chains.parity.io:443`
- [x] `wss://westend-try-runtime-node.parity-chains.parity.io:443`