@ggwpez 2023-05-02
Tested it for all four parity nodes and it works with this patch, thanks!