@liamaharon 2023-05-03
I got quite far through refactoring the code to use `OverlayedChanges`, however it turned into quite a large, non-trivial change due to many areas in the codepath where overlay changes were flushed. 

I opted instead to save/restore the raw trie directly at the `TrieBackend` layer, providing a similar performance improvement without the issue of needing to be really careful to not accidentally flush the overlay somewhere, which causes a blockage of the main thread for many seconds.