@ggwpez 2023-05-04
> With the end goal of making this usable (in a stable and fast way ) in our CI, it would make a lot of sense to have snapshots ready in our infra, and have the CI use those. With this fast loading time, it would be a smooth process.

It is pretty fast now. Takes about 3 minutes locally and 15 in CI.  
I think we should rather focus on stability and add tests instead of optimizing more.