@liamaharon 2023-05-04
@ggwpez `utils/frame/remote-externalities/test_data/proxy_test` is a snapshot file used in a test. I needed to update it to use the new snapshot format.